requirements.txt
annotated-types==0.7.0
anyio==4.5.2
astor==0.8.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.9.1
numpy==1.19.3
openai==2.2.0
pillow==10.4.0
protobuf==5.29.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing_extensions==4.13.2
urllib3==2.2.3
